feat: report to Symon everytime notifications triggered
