+++ src/components/probe/index.ts
@@ -22,15 +22,15 @@
  * SOFTWARE.                                                                      *
  **********************************************************************************/
 
-import { ValidateResponseStatus } from '../notification/alert'
+import { ValidateResponseStatus, validateResponse } from '../notification/alert'
 import { processProbeStatus } from '../notification/process-server-status'
 import { probing } from './probing'
-import { validateResponse } from '../notification/alert'
 import { Probe } from '../../interfaces/probe'
 import { Notification } from '../../interfaces/notification'
 import { notificationLog, probeLog } from '../logger'
 import { AxiosResponseWithExtraData } from '../../interfaces/request'
 import { sendAlerts } from '../notification'
+import { getLogsAndReport } from '../reporter'
 
 /**
  * doProbe sends out the http request
@@ -90,19 +90,22 @@ export async function doProbe(
         notifications &&
         notifications?.length > 0
       ) {
-        notifications.forEach((notification) => {
-          notificationLog({
-            probe,
-            notification,
-            type:
-              status.state === 'UP_TRUE_EQUALS_THRESHOLD'
-                ? 'NOTIFY-INCIDENT'
-                : 'NOTIFY-RECOVER',
-            alertType: probe.alerts[index],
-            url: probe.requests[requestIndex].url,
+        const notificationsLogPromise = Promise.all(
+          notifications.map((notification) => {
+            return notificationLog({
+              probe,
+              notification,
+              type:
+                status.state === 'UP_TRUE_EQUALS_THRESHOLD'
+                  ? 'NOTIFY-INCIDENT'
+                  : 'NOTIFY-RECOVER',
+              alertType: probe.alerts[index],
+              url: probe.requests[requestIndex].url,
+            })
           })
-        })
-        await sendAlerts({
+        )
+
+        const sendAlertsPromise = sendAlerts({
           validation: validatedResp[index],
           notifications: notifications,
           url: probe.requests[requestIndex].url ?? '',
@@ -113,6 +116,11 @@ export async function doProbe(
           statusCode: probeRes.status,
           responseTime: probeRes.config.extraData?.responseTime as number,
         })
+
+        await notificationsLogPromise
+        await sendAlertsPromise
+
+        getLogsAndReport()
       }
     })
   } catch (error) {
